Updated package structure. Clean up some part of the documentation.
