--- docs/PyRCN/echo_state_network/BaseEchoStateNetwork.rst
@@ -1,5 +1,0 @@
-pyrcn.echo_state_network.BaseEchoStateNetwork
-=============================================
-
-.. automodule:: pyrcn.echo_state_network.BaseEchoStateNetwork
-   :members:
